created views package and separated view classes
