+++ inventory_requests/urls.py
@@ -1,10 +1,13 @@
 from django.conf.urls import url
-from inventory_requests import views
-
+from inventory_requests.views.RequestList import RequestList
+from inventory_requests.views.DetailedRequest import ViewDetailedRequest
+from inventory_requests.views.RequestListUser import RequestListUser
+from inventory_requests.views.CreateRequest import CreateRequest
+from inventory_requests.views.ModifyRequest import ModifyRequest
 urlpatterns = [
-    url(r'^requests/$', views.RequestList.as_view()),
-    url(r'^requests/(?P<pk>[0-9]+)/$', views.ViewDetailedRequest.as_view()),
-    url(r'^userRequests', views.RequestListUser.as_view()),
-    url(r'^createRequest', views.CreateRequest.as_view()),
-    url(r'^modifyRequest/(?P<pk>[0-9]+)/$', views.ModifyRequest.as_view()),
+    url(r'^requests/$', RequestList.as_view(), name='requests-list'),
+    url(r'^requests/(?P<pk>[0-9]+)/$', ViewDetailedRequest.as_view(), name='detailed-request'),
+    url(r'^userRequests', RequestListUser.as_view(), name='user-requests'),
+    url(r'^createRequest', CreateRequest.as_view(), name='create-request'),
+    url(r'^modifyRequest/(?P<pk>[0-9]+)/$', ModifyRequest.as_view(), name='modify-request'),
 ]
